deixa modais genericos
