+++ src/components/Shareable/ModalCODAEQuestionaFinalForm/index.jsx
@@ -1,23 +1,17 @@
-import React, { useContext, useState } from "react";
+import React from "react";
 import { Field, Form } from "react-final-form";
-import { OnChange } from "react-final-form-listeners";
 import { Modal } from "react-bootstrap";
 import HTTP_STATUS from "http-status-codes";
-import { composeValidators, getError } from "helpers/utilities";
-import {
-  peloMenosUmCaractere,
-  textAreaRequired
-} from "helpers/fieldValidators";
+import { getError } from "helpers/utilities";
 import Botao from "components/Shareable/Botao";
 import {
   BUTTON_STYLE,
   BUTTON_TYPE
 } from "components/Shareable/Botao/constants";
 import { toastError, toastSuccess } from "components/Shareable/Toast/dialogs";
-import { SolicitacaoAlimentacaoContext } from "context/SolicitacaoAlimentacao";
 import CKEditorField from "components/Shareable/CKEditorField";
 
-export const ModalNegarKitLancheCEMEI = ({ ...props }) => {
+export const ModalCODAEQuestionaFinalForm = ({ ...props }) => {
   const {
     showModal,
     closeModal,
@@ -26,35 +20,25 @@ export const ModalNegarKitLancheCEMEI = ({ ...props }) => {
     loadSolicitacao,
     tipoSolicitacao
   } = props;
-  const [justificativa, setJustificativa] = useState("");
-
-  const solicitacaoAlimentacaoContext = useContext(
-    SolicitacaoAlimentacaoContext
-  );
 
   const onSubmit = async values => {
     const resp = await endpoint(solicitacao.uuid, values, tipoSolicitacao);
     if (resp.status === HTTP_STATUS.OK) {
       closeModal();
-      toastSuccess("Solicitação negada com sucesso!");
-      if (loadSolicitacao) {
-        const response = await loadSolicitacao(solicitacao.uuid);
-        if (response && response.status === HTTP_STATUS.OK) {
-          solicitacaoAlimentacaoContext.updateSolicitacaoAlimentacao(
-            response.data
-          );
-        }
-      }
+      toastSuccess("Questionamento enviado com sucesso!");
+      if (loadSolicitacao) loadSolicitacao();
     } else {
       closeModal();
-      toastError(`Houve um erro ao negar solicitação: ${getError(resp.data)}`);
+      toastError(
+        `Houve um erro ao enviar o questionamento: ${getError(resp.data)}`
+      );
     }
   };
 
   return (
     <Modal dialogClassName="modal-90w" show={showModal} onHide={closeModal}>
       <Modal.Header closeButton>
-        <Modal.Title>Deseja negar a solicitação?</Modal.Title>
+        <Modal.Title>Questionamento</Modal.Title>
       </Modal.Header>
       <Form
         onSubmit={onSubmit}
@@ -62,37 +46,37 @@ export const ModalNegarKitLancheCEMEI = ({ ...props }) => {
         render={({ handleSubmit }) => (
           <form onSubmit={handleSubmit}>
             <Modal.Body>
+              <div className="row">
+                <div className="col-12">
+                  <p className="title">
+                    É possível atender a solicitação com todos os itens
+                    previstos no contrato?
+                  </p>
+                </div>
+              </div>
               <div className="form-row">
                 <div className="form-group col-12">
                   <Field
                     component={CKEditorField}
-                    label="Justificativa"
-                    name="justificativa"
-                    required
-                    validate={composeValidators(
-                      textAreaRequired,
-                      peloMenosUmCaractere
-                    )}
+                    label="Observação"
+                    name="observacao_questionamento_codae"
+                    placeholder="Alguma observação para a Terceirizada?"
                   />
-                  <OnChange name="justificativa">
-                    {value => setJustificativa(value)}
-                  </OnChange>
                 </div>
               </div>
             </Modal.Body>
             <Modal.Footer>
               <Botao
-                texto="Não"
+                texto="Cancelar"
                 type={BUTTON_TYPE.BUTTON}
                 onClick={closeModal}
                 style={BUTTON_STYLE.GREEN_OUTLINE}
                 className="ml-3"
               />
               <Botao
-                texto="Sim"
+                texto="Enviar"
                 type={BUTTON_TYPE.SUBMIT}
                 style={BUTTON_STYLE.GREEN}
-                disabled={justificativa === "" || justificativa === undefined}
                 className="ml-3"
               />
             </Modal.Footer>
